Fix bad import graph from opts/opts.go
